requirements.txt
fastapi[standard]
uvicorn
python-dotenv
openai
openai-agents
PyPDF2